requirements.txt
mss
numpy
opencv-python
Pillow
scikit-image